Use --force-rf64-format when decoding FLACs.
FLAC decodes to the standard WAV format by default, and makes no
consideration when a file requires more than 4GiB to do so, instead
exiting with an error.  FLAC supports two alternative formats to
workaround the limitation, rf64 and wave64.  I picked rf64 somewhat
arbitrarily, and also because ffmpeg considers it an extension of WAV.

For CD quality audio, which is 16-bit 441000 Hz stereo, this would
require a FLAC over 6 hours and 45 minutes to run into the issue.
However, it does rapidly become an issue with greator sample sizes,
rates, and channels.  If you rip a Blu-ray Disc, for example, with a
7.1 soundtrack at 24-bit 192 kHz (not atypical), the uncompressed
audio would exceed 4GiB in just under 16 minutes.

Relatively rare limitation to bump into, but here’s the fix. 😀