No more intermediate decompression and metaflac passes
flac supports reading in a FLAC file directly for source input, and
writes out a new file with the desired compression file based on it.
Use this to speed up operation and make it a little more reliable,
copying all metadata and ancillary streams (such as artwork) without
worry about leaving them out.

Additionally, the `--verify` flag is now used too, giving extra
assurance that the output FLAC file is correct.  I have strong doubts
this is actually necessary; I think decoding errors will be caught and
should cover enough, but this has been requested twice now and the
performance impact is offset by the overall improvement in this
commit.

Thanks @Entze and @wlhlm for the suggestions

Closes: #5, #6